api/requirements.txt
django
psycopg2-binary
djangorestframework
django-filter
numpy
pandas
matplotlib
pytrends
fbprophet
alpha_vantage